pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dl-obs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.25.1",
    "azure-storage-blob>=12.28.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
    "polars>=1.38.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "torch>=2.10.0",
]

[dependency-groups]
dev = [
    "matplotlib>=3.10.8",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.0",
    "seaborn>=0.13.2",
    "types-requests>=2.32.4.20260107",
]

# ============================================================================
# HATCH - Build configuration
# ============================================================================
[tool.hatch.build.targets.wheel]
packages = ["src", "configs", "scripts", "utils", "experiments"]

# ============================================================================
# RUFF - Linter & Formatter (replaces flake8, isort, black)
# ============================================================================
[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (géré par formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["src", "configs", "scripts", "utils", "experiments"]

# ============================================================================
# MYPY - Type checking
# ============================================================================
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

# ============================================================================
# PYTEST - Testing
# ============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]

# ============================================================================
# COVERAGE - Code coverage
# ============================================================================
[tool.coverage.run]
source = ["src", "configs", "scripts", "utils", "experiments"]
branch = true
omit = ["tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
fail_under = 40
show_missing = true
